Add TLS support, self-signed TLS Certs
Makes TLS usable from VMs.

This adds TLS Certificate generation
to Vault, and sets the CA
Environment Variable for Vault and
the corresponding CA Cert Environment
Variable for cURL.

These are self-signed, so they won't
totally validate outside of Vault.

But they work.

I was not able to get the OS Root Trust
Store to update properly, and some of that
code may need to be modified.

This commit also updates documentation to use HTTPS.

Updated the main docs
to show HTTPS for the Vault
servers.

People will need to temporarily
trust the self-signed certificates
generated by Vagrant for these Vault
servers.

Some of the other docs had
http://127.0.0.1:8200 as their
target for demonstrations.

Most of them refer to $VAULT_ADDR,
but I have changed the ones that
refer to a specific IP address.

Update README.md to note Trusting Self-Signed Certificates is needed.